On the issue of rotating the Screen.
What remains is:  Make separate method to cut down on Redundant Code...

And: When you run the timer and stop it.  It shows you the elapsed time.  If you then rotate it, it erases the elapsed time.  This is because when you rotate it it stores the values for hour min sec.  But in the OnCreate method you initialize the number to zero, so the numbers should be initialized to whatever hour min sec is.  Meaning in the very beginning, (first run) it should set hour min sec to 0?  Or to Current time?  What is the default value for the textView?  If you omit the default values for the textView, then the time should be shown correctly.